rename update -> merge
